Improve default configuration.
In particular, it was too easy to have all clients on the same warehouse, as
the number of clients had to be provided explicitly. It is now computed from
the number of warehouses and the number of clients for each warehouse.

The default for the think time is now ON, to cope with the higher numebr of
clients.